add initial version of block navigation buttons
There are 3 navigational elements: go to previous, current or a specific block.
The previous/current block navigation elements are just buttons,
while the go-to-a-specific block element consists of a text field and a button.

Still have a bug related to the go-to-a-specific block that needs to be fixed,
as that element doesn't work for now.